+++ ChangeLog.md
@@ -1,3 +1,7 @@
+2018-11-12
+
+- update missionDef description
+
 2018-11-09
 
 - add page for parameter